yoav: battery design
